refactor: move the status lines into display helpers, return the mtime-sync error, relocate the stream tests, add an ignored --list e2e
Continuing the audit pass: keep the styled terminal output in display.rs,
stop the mtime sync from printing inside the file logic, move the
download_file_content behaviour tests next to the code they test, and give
--list a live end-to-end check.

Terminal output:
- display.rs: eight named print helpers for the file-block status lines
  (retry notice, retry wait, max retries exceeded, stale-file
  re-download, complete-part verification, re-download from scratch, the
  mtime warning, the interrupted-batch line); the formats are verbatim
  from the old inline println!s.
- retry.rs: RetryTracker::record delegates to the display helpers instead
  of composing the three coloured lines itself; the module no longer
  imports colored.
- downloader/mod.rs: the four remaining inline status lines
  (Partial / Resume / Retry / Download-interrupted) use the helpers too.
- mtime.rs: sync_file_mtime returns std::io::Result<bool> instead of
  printing its own warning; the three call sites (install_downloaded_file,
  handle_existing_file, save_xml_metadata) warn through
  print_mtime_warning and keep the batch going exactly as before.

XML metadata:
- archive_metadata.rs: fetch_and_parse_xml sheds the spinner parameter —
  run() sets the "Parsing archive metadata..." message before the fetch,
  so the network module no longer repaints the caller's spinner.
- archive_metadata.rs: xml_file_name_of stops expecting — a Url path
  always splits into at least one segment, so the last one is taken with
  an empty fallback instead of a panic.

Test layout:
- the 14 download_file_content behaviour tests (resume, 206/416, retry,
  empty bodies, mid-stream interrupts) move from the batch module into
  stream.rs, where the code they test lives; the batch and 416-part tests
  stay in downloader/mod.rs.
- the shared fixtures (task, file_task, file_server, ok_empty, fast_retry,
  run_download) move to test_support; prepare_file_for_download and
  download_file_content become pub(crate) for them.

New:
- tests/list.rs: an end-to-end --list run against live archive.org,
  #[ignore] by default (cargo test --test list -- --ignored). It asserts
  only the stdout rows: the "Archive has" banner is the spinner's final
  line, which indicatif draws to stderr only when it is a TTY, so it is
  absent from a piped run.
- AGENTS.md: the architecture listing follows (display's status lines,
  test_support's scripted local mock, plan's structured warnings, the new
  tests/ tree), and the Testing section gains the e2e command.
- README.md: the Features section is rewritten into short bullets and
  re-checked against the current pipeline; --list and --cookies get their
  own entries.

No behaviour change: the on-screen output is unchanged except that the
"Parsing archive metadata..." spinner message now also covers the brief
HEAD pre-check, and the test suite stays at 103 passing (fmt and clippy
-D warnings clean).
